Integrate dataset and flwr

+++ pytorchexample/models/xception.py
@@ -207,4 +207,4 @@ def xception(**kwargs):
     
     #if pretrained:
     #    model.load_state_dict(model_zoo.load_url(model_urls['xception']))
-    return model.double()
+    return model
